--- libs/ui/src/components/atoms/ButtonV2/index.ts
@@ -1,1 +1,0 @@
-export * from './ButtonV2';
