+++ libs/ui/src/components/molecules/CollapsibleCard/CollapsibleCard.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { H4, ParSm } from '../../atoms';
-import { ButtonV2 } from '../../atoms/ButtonV2';
+import { Button } from '../../atoms/Button';
 import { WrappedCheckbox } from '../WrappedCheckbox';
 import { WrappedInput } from '../WrappedInput';
 import { CollapsibleCard } from './CollapsibleCard';
@@ -63,9 +63,9 @@ ProposalCardExample.args = {
           },
         ]}
       />
-      <ButtonV2 variant="outline" colorVariant="secondary">
+      <Button variant="outline" colorVariant="secondary">
         Click Me
-      </ButtonV2>
+      </Button>
     </div>
   ),
   collapsibleContent: (
